refactor: rename migration transaction methods for consistency and add new transaction handlers
