+++ sdk/sdk-server/src/SDKAppRouter.ts
@@ -15,14 +15,14 @@ import { setTipRate } from './armada-protocol-handlers/governance/setTipRate'
 import { updateRebalanceCooldown } from './armada-protocol-handlers/governance/updateRebalanceCooldown'
 import { adjustBuffer } from './armada-protocol-handlers/keepers/adjustBuffer'
 import { rebalance } from './armada-protocol-handlers/keepers/rebalance'
-import { getDepositTx } from './armada-protocol-handlers/users/getDepositTx'
+import { getDepositTx } from './armada-protocol-handlers/users/getDepositTxTx'
 import { getVaultsRaw } from './armada-protocol-handlers/users/getVaults'
 import { getVaultRaw } from './armada-protocol-handlers/users/getVault'
 import { getPoolInfo } from './armada-protocol-handlers/users/getPoolInfo'
 import { getPosition as getArmadaPosition } from './armada-protocol-handlers/users/getPosition'
 import { getUserPositions } from './armada-protocol-handlers/users/getUserPositions'
 import { getUserPosition } from './armada-protocol-handlers/users/getUserPosition'
-import { getWithdrawTx } from './armada-protocol-handlers/users/getWithdrawTx'
+import { getWithdrawTx } from './armada-protocol-handlers/users/getWithdrawTxTx'
 import { buildOrder } from './handlers/buildOrder'
 import { getArmadaSimulation } from './handlers/getArmadaSimulation'
 import { getImportSimulation } from './handlers/getImportSimulation'
